Combines achievementloader and problemloader into loader Updates server.ts to use Loader
